Shrink the fork's footprint on upstream files to 31 lines
Same behaviour; the extensibility code moves off files McNeel owns and into
our own, so future merges stay near-trivial.

- McpDispatcher and RhMcpPlugin become partial. AllTools, TryResolveTool, the
  extension-host field and the GetPlugInObject override — with their docs —
  move to Server/Extensibility/{McpDispatcher,RhMcpPlugin}.Extensibility.cs.
  Those methods and their documentation were the bulk of the delta: adding
  them to upstream files cost 72 lines in McpEndpoint.cs and 25 in Plugin.cs.
- Revert RhMcp.csproj to upstream's 8.29 RhinoCommon/Grasshopper pins. The
  bump to 8.31 existed to resolve CS1705 skew against our engine assembly back
  when the fork project-referenced it; that reference went away when the bridge
  became a pipe client, and the build is green on 8.29.

Per-file, ignoring whitespace: ParameterBinder 19, McpEndpoint 6, Plugin 2,
ToolRegistry 2, manifest.yml 2. Everything else is in files upstream does not
have.

ParameterBinder is the one that should not be here: it fixes an upstream bug
where an async Task<T> tool result is silently dropped, because
AsyncStateMachineBox<TStateMachine,TResult> derives from Task<TResult> rather
than being one, so GetGenericTypeDefinition() never matches. Worth a PR to
McNeel — that would take the fork to 12 lines.

Build: 0 errors.
